Homogenise the storage and querying of bin numbers
Addition of other binning routines means the output of the binning must
be the same regardless of the method used. Previous attributes such as
``vor_output`` and the list of `bin_nums`` have been replaced by a
dictionary also called ``bin_nums``. The format of this dict, to be
populated by the binning routine is:

    self.bin_nums = {[bin_number]:{"spax":(x_spaxels, y_spaxels),
    "mean":(x_mean, y_mean)}, ...}

For voronoi binning, x/y_mean are the centre of mass of the bins whereas
for HII explorer this is the peak.

Custom bins are added as negative bin numbers starting at -1, as
before. The current list of (good) bin numbers is accessible via self._get_bin_nums().