Force emission line bins to be contiguous with minimum size
Previously, within each cutout around a peak, any pixel could be a member of
the bin if it satisfied the conditions of the HIIexplorer algorithm.  Now we
require that it is connected to the peak by pixels that also satisfy the
conditions. Furthermore the number of pixels must be greater than or equal to
``min_npix`` for a bin to be added. These changes act to 'clean up' the bin map
and reduces the number of spurious pixels being included.