Update pickle method
The pickling now stores the FITS data separately from the pickle file as this was causing problems with larger cubes.

The PrimaryHDU from the original FITS file is required for astropy to save a
FITS file. As a result the call to IFUCube now simply takes a HDUList of
[primary, science, stddev] HDUs. MUSECube has been updated to reflect these
changes in the call. Reddening and redshift header entires are stored in this
primary header now.

A clobber option to save_pkl() has been added.