Add border arg to emission_line_bin
Previously, any peaks found by emission_line_bin within `r` of the cube edge
were discarded. This didn't work correctly in the regime of cubes not
orientated NE since the cube is padded. Now this has been decoupled from the
value of `r` (the max radius of the bins) and now `border` can be used to
specify the minimum distance to a nan value (or the cube edge) for a peak to be
valid. Note this will also exclude peaks within `border` of masked regions as
they are replaced by nans.