Add Balmer decrement calculation
The Balmer decrement (from the fiducial 2.86 value) is calculated to give an
E(B-V)_gas. This is used to deredden the fluxes of the emission lines. Currently the uncertainty in E(B-V)_gas is not propagated onto the flux uncertainty.

get_Alamb has been updated also to return k_lambda [i.e. A_lambda/E(B-V)] to
facilitate this.