Add option to choose starlight tmp directory
By default, run_starlight() will put output in the system's tmp location. It
can now be manually set with the ``tmp_dir`` argument. This is in order to help
in the cases of long file paths to the tmp location, which can cause starlight
to fail.

Additionally, to aid long file path prevention, the `starlight_` prefix has
been changed to `sl_`